:twisted_rightwards_arrows: Merge(feature/stock Branch ->dev Branch): stock api 예외 처리 추가, 인기,관심 종목 응답dto수정

+++ src/main/java/com/heartfoilo/demo/domain/invest/entity/Order.java
@@ -31,7 +31,7 @@ public class Order {
     private LocalDateTime orderDate;
 
     @Column(name = "order_amount", nullable = false)
-    private int orderAmount;
+    private Long orderAmount;
 
     @Column(name = "order_price", nullable = false)
     private int orderPrice;
